Rename tangram.h/cpp to map.h/cpp
This header encompasses the Map class, and should be named accordingly.
