+++ bench/src/tileLoading.cpp
@@ -1,8 +1,8 @@
-#include "tangram.h"
-#include "gl.h"
-#include "platform_mock.h"
-#include "log.h"
 #include "data/tileSource.h"
+#include "gl.h"
+#include "log.h"
+#include "map.h"
+#include "platform_mock.h"
 #include "scene/sceneLoader.h"
 #include "scene/scene.h"
 #include "style/style.h"
