+++ core/src/gl/texture.cpp
@@ -3,10 +3,10 @@
 #include "gl/glError.h"
 #include "gl/renderState.h"
 #include "gl/hardware.h"
-#include "util/geom.h"
 #include "log.h"
+#include "map.h"
 #include "platform.h"
-#include "tangram.h"
+#include "util/geom.h"
 
 // Enable only JPEG, PNG, GIF, TGA and PSD
 #define STBI_NO_BMP
