+++ core/src/labels/labels.cpp
@@ -7,13 +7,13 @@
 #include "labels/labelSet.h"
 #include "labels/obbBuffer.h"
 #include "labels/textLabel.h"
+#include "map.h"
 #include "marker/marker.h"
 #include "platform.h"
 #include "scene/scene.h"
 #include "style/pointStyle.h"
 #include "style/style.h"
 #include "style/textStyle.h"
-#include "tangram.h"
 #include "tile/tile.h"
 #include "tile/tileCache.h"
 #include "tile/tileManager.h"
