+++ core/src/style/pointStyleBuilder.cpp
@@ -1,15 +1,15 @@
 #include "style/pointStyleBuilder.h"
 
 #include "data/propertyItem.h"
-#include "marker/marker.h"
 #include "labels/labelCollider.h"
 #include "labels/spriteLabel.h"
 #include "log.h"
+#include "marker/marker.h"
+#include "map.h"
 #include "scene/drawRule.h"
 #include "scene/spriteAtlas.h"
 #include "scene/stops.h"
 #include "selection/featureSelection.h"
-#include "tangram.h"
 #include "tile/tile.h"
 #include "util/geom.h"
 #include "util/lineSampler.h"
