+++ core/src/style/polylineStyle.cpp
@@ -5,12 +5,12 @@
 #include "gl/texture.h"
 #include "gl/renderState.h"
 #include "log.h"
+#include "map.h"
 #include "marker/marker.h"
 #include "material.h"
 #include "platform.h"
 #include "scene/stops.h"
 #include "scene/drawRule.h"
-#include "tangram.h"
 #include "tile/tile.h"
 #include "util/builders.h"
 #include "util/dashArray.h"
