+++ core/src/style/style.cpp
@@ -5,13 +5,13 @@
 #include "gl/shaderProgram.h"
 #include "gl/mesh.h"
 #include "log.h"
+#include "map.h"
 #include "marker/marker.h"
 #include "scene/light.h"
 #include "scene/scene.h"
 #include "scene/spriteAtlas.h"
 #include "scene/styleParam.h"
 #include "style/material.h"
-#include "tangram.h"
 #include "tile/tile.h"
 #include "view/view.h"
