+++ core/src/style/textStyleBuilder.cpp
@@ -6,10 +6,10 @@
 #include "labels/textLabel.h"
 #include "labels/textLabels.h"
 #include "log.h"
+#include "map.h"
 #include "marker/marker.h"
 #include "selection/featureSelection.h"
 #include "scene/drawRule.h"
-#include "tangram.h"
 #include "tile/tile.h"
 #include "util/geom.h"
 #include "util/mapProjection.h"
