+++ core/src/tile/tileWorker.cpp
@@ -2,8 +2,8 @@
 
 #include "data/tileSource.h"
 #include "log.h"
+#include "map.h"
 #include "platform.h"
-#include "tangram.h"
 #include "tile/tileBuilder.h"
 #include "tile/tileID.h"
 #include "tile/tileTask.h"
